+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Change Bottles to be installed as a Flatpak.
 - Change Steam to be installed as part of the first-time setup.
 - Add documentation in the read me file for doing customized installations on bare-metal.
+- Change the Minecraft launcher from MultiMC to PolyMC and have it installed as a Flatpak.
 
 ## 3.0.0